+++ crates/mimic_tests/src/lib.rs
@@ -4,16 +4,15 @@
 
 #[cfg(test)]
 pub mod test {
-    use icu::{Log, log};
+    use test_design::fixture::rarity::Rarity;
 
     #[test]
     fn test_icu_crate() {
-        log!(Log::Info, "icu v{}", icu::VERSION);
-        log!(Log::Info, "mimic v{}", mimic::VERSION);
-        assert!(!icu::VERSION.is_empty(), "icu VERSION should be non-empty");
-        assert!(
-            !mimic::VERSION.is_empty(),
-            "mimic VERSION should be non-empty"
-        );
+        let r = Rarity::default();
+
+        let errs = mimic::core::validate(&r);
+        if let Err(e) = errs {
+            println!("{e}");
+        }
     }
 }